# swyftr-swagger.yaml
swagger: '2.0'
info:
  version: '1'
  title: Swyftr
  description: Backend API for Swyftr system
host: 52.221.58.5
basePath: /api/v1
schemes:
  - https
...
paths:
...
  /store:
    post:
      summary: Create Store
      tags:
        - Sprint2
        - Store
        - System
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/Store'
      responses:
        200:
          description: OK
          schema:
            type: object
            properties:
              status:
                type: integer
              storeId:
                type: integer
        '400':
          description: Bad Request
        '500':
          description: Fail
    patch:
      summary: Update Store
      tags:
        - Sprint2
        - Store
        - System
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/Store'
      responses:
        200:
          description: OK
        '400':
          description: Bad Request
        '500':
          description: Fail
    get:
      summary: Get store by store id
      tags:
        - Sprint2
        - Store
        - System
      produces:
        - application/json
      parameters:
        - in: query
          name: storeId
          required: true
          type: integer
      responses:
        200:
          description: OK
          schema:
            $ref: '#/definitions/Store'
        '400':
          description: Bad Request
        '500':
          description: Fail
  /store/cus/:
    get:
      summary: Get store by store id
      tags:
        - Sprint2
        - Store
        - System
      produces:
        - application/json
      parameters:
        - in: query
          name: storeId
          required: true
          type: integer
      responses:
        200:
          description: OK
          schema:
            $ref: '#/definitions/Store'
        '400':
          description: Bad Request
        '500':
          description: Fail
...
  /store/all/{partnerId}:
    get:
      summary: Get all stores belong to a partner
      tags:
        - Sprint2
        - Store
        - System
      produces:
        - application/json
      parameters:
        - in: path
          name: partnerId
          type: string
          required: true
      responses:
        200:
          description: OK
          schema:
            type: object
            properties:
              status:
                type: integer
              stores:
                type: array
                items:
                  $ref: '#/definitions/Store'
  /store/allProducts/{storeId}:
    get:
      summary: Get all products belongs to a store
      tags:
        - Sprint2
        - Store
        - Product
        - System
      produces:
        - application/json
      parameters:
        - in: path
          name: storeId
          type: number
          required: true
      responses:
        200:
          description: OK
          schema:
            type: object
            properties:
              status:
                type: integer
              products:
                type: array
                items:
                  $ref: '#/definitions/Product'
              store:
                $ref: '#/definitions/Store'
  /store/cus/all/{partnerId}:
    get:
      summary: Get all stores/branches for an organization
      tags:
        - Store
        - Customer
        - Sprint2
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: path
          name: partnerId
          type: number
          required: true
      responses:
        200:
          description: OK
          schema:
            type: object
            properties:
              status:
                type: number
              stores:
                type: array
                items:
                  $ref: '#/definitions/Store'
        400:
          description: Bad Request
        500:
          description: Server Error
        401:
          description: Unauthorized
  /store/cus/allProducts/{storeId}:
    get:
      summary: Get all products that are allowed to sale in a store
      tags:
        - Sprint2
        - Customer
        - Store
      produces:
        - application/json
      parameters:
        - in: path
          name: storeId
          required: true
          type: number
      responses:
        200:
          description: OK
          schema:
            type: object
            properties:
              status:
                type: number
              products:
                type: array
                items:
                  $ref: '#/definitions/Product'
              store:
                $ref: '#/definitions/Store'
        400:
          description: Bad Request
        500:
          description: Server Error
        401:
          description: Unauthorized
...
  /report/settlements/store/{status}/{fromDate}/{toDate}/{branchId}:
    get:
      summary: Settlements By Store
      tags:
        - Sprint2
        - Reporting
        - Store
      produces:
        - application/json
      parameters:
        - in: path
          name: status
          type: string
          required: true
        - in: path
          name: fromDate
          type: string
          required: true
        - in: path
          name: toDate
          type: string
          required: true
        - in: path
          name: branchId
          type: integer
          required: true
      responses:
        200:
          description: OK
          schema:
            type: object
            properties:
              status:
                type: integer
              orders:
                type: array
                items:
                  $ref: '#/definitions/Settlements_Response'
...
  /report/items/store/{status}/{fromDate}/{toDate}/{branchId}:
    get:
      summary: Number Of Items sold per Store
      tags:
        - Sprint2
        - Reporting
        - Store
      produces:
        - application/json
      parameters:
        - in: path
          name: status
          type: string
          required: true
        - in: path
          name: fromDate
          type: string
          required: true
        - in: path
          name: toDate
          type: string
          required: true
        - in: path
          name: branchId
          type: integer
          required: true
      responses:
        200:
          description: OK
          schema:
            type: object
            properties:
              status:
                type: integer
              orders:
                type: array
                items:
                  $ref: '#/definitions/Items_Sold_Response'
...
  /report/orders/store/{status}/{fromDate}/{toDate}/{branchId}:
    get:
      summary: Orders By Status By Store
      tags:
        - Sprint2
        - Reporting
        - Store
      produces:
        - application/json
      parameters:
        - in: path
          name: status
          type: string
          required: true
        - in: path
          name: fromDate
          type: string
          required: true
        - in: path
          name: toDate
          type: string
          required: true
        - in: path
          name: branchId
          type: integer
          required: true
      responses:
        200:
          description: OK
          schema:
            type: object
            properties:
              status:
                type: integer
              orders:
                type: array
                items:
                  $ref: '#/definitions/Order_Status_Store'
...
definitions:
...
  Store:
    type: object
    properties:
      id:
        type: integer
      orgId:
        type: integer
      branchName:
        type: string
      address:
        type: string
      longitude:
        type: number
      latitude:
        type: number
      district:
        type: string
      province:
        type: string
      managerId:
        type: integer
      phone:
        type: string
      logo:
        type: number
      displayImage:
        type: number
      slogan:
        type: string
      branchLogoImage:
        type: string
      branchDisplayImage:
        type: string
...
  Product:
    type: object
    properties:
      id:
        type: integer
      productName:
        type: string
      productCode:
        type: string
      qty:
        type: number
      unitPrice:
        type: number
      branchId:
        type: number
      categoryId:
        type: number
      description:
        type: string
      discount:
        type: number
      availability:
        type: string
        enum: ['IN', 'OUT']
      specialNote:
        type: string
      assetIds:
        type: array
        items:
          type: integer      
      tags:
        type: string
...
  Settlements_Response:
    type: object
    properties:
      orgId:
        type: integer
      orgName:
        type: string
      total:
        type: number
      paymentPercentage:
        type: number
      branchId:
        type: integer
      branchName:
        type: string
  Items_Sold_Response:
    type: object
    properties:
      orgId:
        type: integer
      name:
        type: string
      branchStockId:
        type: integer
      total:
        type: number
      branchId:
        type: integer
      branchName:
        type: string
...
  Order_Status_Store:
    type: object
    properties:
      name:
        type: string
      total:
        type: number
      branchId:
        type: integer
      branchName:
        type: string
...